Initial commit : tech interview for QA automation engineer role
Based on the requirements below we need you to come up with a very basic plan to test the Grammarly Chrome extension. We would also like you to automate the test cases you find using a framework of your choice. You can find Grammarly here.

Requirements:

Test if the extension is enabled or not.
One positive test case - the extension is working as expected.
One negative test case - the extension is not working as expected.
You can consider any use case/tool to test this extension on.

Please respond back with the following outputs:

Create a simple test plan to check the behavior of this extension, according to the requirements above. You can mention the test cases in an excel sheet and mention your automation strategy.
Create a script in any tool you are familiar with to automate these 3 test cases. Provide us with the script and the instructions on how to run it.
Add a test report of your test run.
Add a screencast of your test suite running.